Broaden unsigned coordinate cleanup
Agent-Logs-Url: https://github.com/KRcpl88/WinAmy4D/sessions/e9379f74-9bbb-4c64-8a9b-4e9bf3196ff2

+++ include/bitboard.h
@@ -52,8 +52,8 @@ class CBitBoard {
     void ClearLowestBit() { m_ullBits &= m_ullBits - 1; }
 
     // Static mask constructors
-    static CBitBoard SetMask(int i) { return CBitBoard(1ULL << i); }
-    static CBitBoard ClrMask(int i) { return CBitBoard(~(1ULL << i)); }
+    static CBitBoard SetMask(uint16_t i) { return CBitBoard(1ULL << i); }
+    static CBitBoard ClrMask(uint16_t i) { return CBitBoard(~(1ULL << i)); }
 
     // Counting and scanning
     int CountBits() const {
